Trying to UV map drawSphere function

+++ Lab9/GraphicsProgramming/GraphicsProgramming/Shape.h
@@ -21,7 +21,7 @@ public:
 	void drawCone(float radius, float edges, float height, float x, float y, float z);
 	void drawDisc(int edges, float radius, float x, float y, float z);
 	void drawFlatDisc(int edges, float radius, float h, float k);
-	void drawSphere(float radius, int lats, int longs, float h, float k);
+	void drawSphere(float radius, float lats, float longs, float h, float k);
 	void drawIcosahedron();
 	void normalize(float v[3]);
 	void normcrossprod(float v1[3], float v2[3], float out[3]);
